requirements.txt
fastapi
uvicorn
faiss-cpu
groq
python-dotenv
numpy
pydantic